Promote GA features from the beta IChatClient to the stable IChatClient
